LayoutTests/inspector/elements/elements-delete-inline-style-expected.txt

Tests that the "style" attribute removal results in the Styles sidebar pane update (not a crash). Bug 51478

Before style property removal:
[expanded] 
element.style  { ()
color: red;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;


After style property removal:
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;
